+++ src/containers/shared/App/IntlWrapper.tsx
@@ -5,7 +5,7 @@ import { Select, LocaleProvider } from 'antd'
 import { Locale } from 'antd/lib/locale-provider'
 
 import styles from './index.scss'
-import { useOnMount } from '@utils/reactExt'
+import { useOnMount } from '@utils/hooks'
 import { setCookie } from '@utils/index'
 import { COOKIE_KEYS } from '@constants/index'
 import PageLoading from '@components/PageLoading'
