+++ src/containers/shared/PrivateRoute/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { observer } from 'mobx-react'
 import { Route, RouteProps } from 'react-router-dom'
 
-import { useOnMount } from '@utils/reactExt'
+import { useOnMount } from '@utils/hooks'
 import useRootStore from '@store/useRootStore'
 
 function PrivateRoute({ component: Component, ...rest }: RouteProps) {
